scripts/requirements.txt
#requests-toolbelt==1.0.0
python-dotenv==1.1.0
doccano-client==1.2.8
urllib3==1.26.15